pyproject.toml
[project]
name = "with-adk-a2a-a2ui-agui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "a2a-sdk>=0.3.22",
    "fastapi>=0.128.0",
    "google-adk>=1.18.0",
    "google-generativeai>=0.8.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
